Extracted the ClipTextArea component
